refactor(dashboard): split god-object into blueprint package + service layer
Convert the 2295-line koan/app/dashboard.py into a Flask blueprint package
(koan/app/dashboard/) built by a create_app() factory mirroring api/__init__.py,
with pure business logic extracted to koan/app/dashboard_service/.

- 7 domain blueprints: core, missions, chat, usage, agent, config, prs
- Patchable globals centralized in dashboard/state.py; gate/filters in _helpers.py
- Service layer (missions/journal/plans/stats) unit-tested without a Flask client
- Templates relocated to koan/templates/dashboard/
- Runnable entry app/dashboard/__main__.py (make dashboard, pid_manager, docker)
- Migrated all dashboard test patch targets; added factory + size-gate + service tests
- __init__.py < 300 lines, every blueprint < 400 lines
