+++ Makefile
@@ -121,7 +121,7 @@ migrate: setup
 	$(KOAN_RUN) app/migrate_memory.py
 
 dashboard: setup
-	$(KOAN_RUN) app/dashboard.py $(if $(KOAN_DASHBOARD_HOST),--host $(KOAN_DASHBOARD_HOST),) $(if $(KOAN_DASHBOARD_PORT),--port $(KOAN_DASHBOARD_PORT),)
+	$(KOAN_RUN) app/dashboard/__main__.py $(if $(KOAN_DASHBOARD_HOST),--host $(KOAN_DASHBOARD_HOST),) $(if $(KOAN_DASHBOARD_PORT),--port $(KOAN_DASHBOARD_PORT),)
 
 koan: setup
 	@$(KOAN_RUN) -m app.koan_cli $(PWD)
