+++ koan/templates/dashboard/agent.html
