+++ koan/templates/dashboard/chat.html
