+++ koan/templates/dashboard/config.html
