+++ koan/templates/dashboard/logs.html
