+++ koan/templates/dashboard/progress.html
