+++ koan/templates/dashboard/skills.html
