+++ koan/tests/test_dashboard.py
@@ -10,8 +10,9 @@ from pathlib import Path
 from unittest.mock import patch, MagicMock
 
 from app import dashboard
+from app import dashboard_service
 
-REAL_TEMPLATES = Path(__file__).parent.parent / "templates"
+REAL_TEMPLATES = Path(__file__).parent.parent / "templates" / "dashboard"
 
 
 @pytest.fixture
@@ -49,15 +50,15 @@ def app_client(instance_dir, tmp_path):
     # Copy real templates so Flask can render them
     tpl_dest = tmp_path / "koan" / "templates"
     shutil.copytree(REAL_TEMPLATES, tpl_dest)
-    with patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-         patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"), \
-         patch.object(dashboard, "OUTBOX_FILE", instance_dir / "outbox.md"), \
-         patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-         patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-         patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"), \
-         patch.object(dashboard, "PENDING_FILE", instance_dir / "journal" / "pending.md"), \
-         patch.object(dashboard, "RECURRING_FILE", instance_dir / "recurring.json"), \
-         patch.object(dashboard, "KOAN_ROOT", tmp_path):
+    with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+         patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"), \
+         patch.object(dashboard.state, "OUTBOX_FILE", instance_dir / "outbox.md"), \
+         patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+         patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+         patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"), \
+         patch.object(dashboard.state, "PENDING_FILE", instance_dir / "journal" / "pending.md"), \
+         patch.object(dashboard.state, "RECURRING_FILE", instance_dir / "recurring.json"), \
+         patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
         dashboard.app.config["TESTING"] = True
         dashboard.app.jinja_loader = FileSystemLoader(str(tpl_dest))
         with dashboard.app.test_client() as client:
@@ -66,15 +67,15 @@ def app_client(instance_dir, tmp_path):
 
 class TestParsingMissions:
     def test_parse_sections(self, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
-            result = dashboard.parse_missions()
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
+            result = dashboard_service.missions.parse_missions()
             assert len(result["pending"]) == 2
             assert len(result["in_progress"]) == 1  # complex block
             assert len(result["done"]) == 1
 
     def test_parse_empty(self, tmp_path):
-        with patch.object(dashboard, "MISSIONS_FILE", tmp_path / "nope.md"):
-            result = dashboard.parse_missions()
+        with patch.object(dashboard.state, "MISSIONS_FILE", tmp_path / "nope.md"):
+            result = dashboard_service.missions.parse_missions()
             assert result == {"pending": [], "in_progress": [], "done": []}
 
 
@@ -104,7 +105,7 @@ class TestRoutes:
         assert b"Build dashboard" in resp.data
 
     def test_add_mission(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/missions/add", data={
                 "mission": "New test mission",
                 "project": "koan",
@@ -114,7 +115,7 @@ class TestRoutes:
             assert "[project:koan] New test mission" in content
 
     def test_add_mission_no_project(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/missions/add", data={
                 "mission": "Simple mission",
                 "project": "",
@@ -148,7 +149,7 @@ class TestRoutes:
         assert data["missions"]["in_progress"] == 1
 
     def test_chat_send_mission(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/chat/send", data={
                 "message": "Do something cool",
                 "mode": "mission",
@@ -584,55 +585,55 @@ class TestUsageApi:
 
 class TestSignals:
     def test_no_signals(self, tmp_path):
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["stop_requested"] is False
             assert status["quota_paused"] is False
             assert status["loop_status"] == ""
 
     def test_stop_signal(self, tmp_path):
         (tmp_path / ".koan-stop").write_text("STOP")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["stop_requested"] is True
 
     def test_loop_status(self, tmp_path):
         (tmp_path / ".koan-status").write_text("3/20")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["loop_status"] == "3/20"
 
 
 class TestJournal:
     def test_get_entries(self, instance_dir):
-        with patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
-            entries = dashboard.get_journal_entries(limit=7)
+        with patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
+            entries = dashboard_service.journal.get_journal_entries(limit=7)
             assert len(entries) == 1
             assert entries[0]["date"] == "2026-02-01"
             assert entries[0]["entries"][0]["project"] == "koan"
 
     def test_empty_journal(self, tmp_path):
-        with patch.object(dashboard, "JOURNAL_DIR", tmp_path / "journal"):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", tmp_path / "journal"):
+            entries = dashboard_service.journal.get_journal_entries()
             assert entries == []
 
 
 class TestChatSend:
     """Test /chat/send endpoint — the Claude chat handler."""
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_success(self, mock_run, mock_fmt, mock_hist, mock_save,
                           mock_tools_desc, mock_tools, app_client, instance_dir):
         mock_run.return_value = MagicMock(stdout="Salut !", returncode=0)
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "hello", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
@@ -640,30 +641,30 @@ class TestChatSend:
         assert data["response"] == "Salut !"
         mock_run.assert_called_once()
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_empty_response_fallback(self, mock_run, mock_fmt, mock_hist, mock_save,
                                           mock_tools_desc, mock_tools, app_client, instance_dir):
         mock_run.return_value = MagicMock(stdout="", returncode=0)
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "hello", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
         assert "Try again?" in data["response"]
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_timeout_lite_retry_succeeds(self, mock_run, mock_fmt, mock_hist, mock_save,
                                                mock_tools_desc, mock_tools, app_client, instance_dir):
         """First call times out, lite retry succeeds."""
@@ -671,42 +672,42 @@ class TestChatSend:
             subprocess.TimeoutExpired("claude", 120),
             MagicMock(stdout="Réponse lite !", returncode=0),
         ]
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "deep question", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
         assert data["response"] == "Réponse lite !"
         assert mock_run.call_count == 2
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_timeout_both_attempts(self, mock_run, mock_fmt, mock_hist, mock_save,
                                          mock_tools_desc, mock_tools, app_client, instance_dir):
         """Both full and lite calls time out."""
         mock_run.side_effect = subprocess.TimeoutExpired("claude", 120)
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "deep question", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
         assert "Timeout" in data["response"]
         assert mock_run.call_count == 2
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_timeout_lite_empty_response(self, mock_run, mock_fmt, mock_hist, mock_save,
                                                mock_tools_desc, mock_tools, app_client, instance_dir):
         """First call times out, lite retry returns empty."""
@@ -714,21 +715,21 @@ class TestChatSend:
             subprocess.TimeoutExpired("claude", 120),
             MagicMock(stdout="", returncode=0),
         ]
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "deep question", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
         assert "Timeout" in data["response"]
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_timeout_lite_retry_error(self, mock_run, mock_fmt, mock_hist, mock_save,
                                             mock_tools_desc, mock_tools, app_client, instance_dir):
         """First call times out, lite retry raises OSError."""
@@ -736,47 +737,47 @@ class TestChatSend:
             subprocess.TimeoutExpired("claude", 120),
             OSError("broken"),
         ]
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "hi", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is False
         assert "broken" in data["error"]
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_exception(self, mock_run, mock_fmt, mock_hist, mock_save,
                             mock_tools_desc, mock_tools, app_client, instance_dir):
         mock_run.side_effect = OSError("claude not found")
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "hi", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is False
         assert "claude not found" in data["error"]
 
-    @patch("app.dashboard.get_allowed_tools", return_value="")
-    @patch("app.dashboard.get_tools_description", return_value="")
-    @patch("app.dashboard.save_conversation_message")
-    @patch("app.dashboard.load_recent_history", return_value=[])
-    @patch("app.dashboard.format_conversation_history", return_value="")
-    @patch("app.dashboard.subprocess.run")
+    @patch("app.dashboard.chat.get_allowed_tools", return_value="")
+    @patch("app.dashboard.chat.get_tools_description", return_value="")
+    @patch("app.dashboard.chat.save_conversation_message")
+    @patch("app.dashboard.chat.load_recent_history", return_value=[])
+    @patch("app.dashboard.chat.format_conversation_history", return_value="")
+    @patch("app.dashboard.chat.subprocess.run")
     def test_chat_empty_response_logs_stderr(self, mock_run, mock_fmt, mock_hist, mock_save,
                                               mock_tools_desc, mock_tools, app_client, instance_dir, capsys):
         """When Claude returns empty stdout with stderr, stderr should be logged."""
         mock_run.return_value = MagicMock(stdout="", stderr="model overloaded", returncode=1)
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.post("/chat/send", data={"message": "hello", "mode": "chat"})
         data = resp.get_json()
         assert data["ok"] is True
@@ -784,7 +785,7 @@ class TestChatSend:
         assert "model overloaded" in captured.out
 
     def test_chat_send_with_project_tag(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/chat/send", data={
                 "message": "[project:koan] add feature",
                 "mode": "mission",
@@ -800,14 +801,14 @@ class TestBuildDashboardPrompt:
     """Test _build_dashboard_prompt lite mode."""
 
     def test_lite_prompt_strips_journal_and_summary(self, instance_dir):
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-             patch("app.dashboard.load_recent_history", return_value=[]), \
-             patch("app.dashboard.format_conversation_history", return_value=""), \
-             patch("app.dashboard.get_tools_description", return_value=""):
-            prompt = dashboard._build_dashboard_prompt("hello", lite=True)
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+             patch("app.dashboard.chat.load_recent_history", return_value=[]), \
+             patch("app.dashboard.chat.format_conversation_history", return_value=""), \
+             patch("app.dashboard.chat.get_tools_description", return_value=""):
+            prompt = dashboard.chat._build_dashboard_prompt("hello", lite=True)
         assert "Session 1: bootstrapped" not in prompt
         assert "Built the dashboard" not in prompt
         assert "You are Kōan" in prompt
@@ -815,17 +816,17 @@ class TestBuildDashboardPrompt:
 
 class TestParseProject:
     def test_english_tag(self):
-        project, text = dashboard.parse_project("[project:koan] fix bug")
+        project, text = dashboard.chat.parse_project("[project:koan] fix bug")
         assert project == "koan"
         assert text == "fix bug"
 
     def test_french_tag(self):
-        project, text = dashboard.parse_project("[projet:koan] fix bug")
+        project, text = dashboard.chat.parse_project("[projet:koan] fix bug")
         assert project == "koan"
         assert text == "fix bug"
 
     def test_no_tag(self):
-        project, text = dashboard.parse_project("fix bug")
+        project, text = dashboard.chat.parse_project("fix bug")
         assert project is None
         assert text == "fix bug"
 
@@ -866,7 +867,7 @@ class TestApiProgressStream:
         # iterating the infinite SSE generator (which blocks ~14s in the test
         # client waiting for output).
         with dashboard.app.test_request_context("/api/progress/stream"):
-            resp = dashboard.api_progress_stream()
+            resp = dashboard.chat.api_progress_stream()
         assert resp.content_type == "text/event-stream; charset=utf-8"
         assert resp.headers.get("Cache-Control") == "no-cache"
         assert resp.headers.get("X-Accel-Buffering") == "no"
@@ -876,7 +877,7 @@ class TestApiProgressStream:
         pending.write_text("# Mission: live test\n---\n04:30 — doing stuff\n")
         # Patch time.sleep to terminate the generator after the first event.
         # The generator yields a data event (file exists) then hits sleep → error.
-        with patch("app.dashboard.time.sleep", side_effect=RuntimeError("break")):
+        with patch("app.dashboard.chat.time.sleep", side_effect=RuntimeError("break")):
             resp = app_client.get("/api/progress/stream")
         # Read the first chunk from the streaming response
         import json
@@ -900,7 +901,7 @@ class TestApiProgressStream:
         # Call the view function directly — we only need to verify the Response
         # metadata, not iterate the generator.
         with dashboard.app.test_request_context("/api/progress/stream"):
-            resp = dashboard.api_progress_stream()
+            resp = dashboard.chat.api_progress_stream()
         assert "text/event-stream" in resp.content_type
 
 
@@ -913,28 +914,28 @@ class TestSignalStatusPause:
 
     def test_pause_signal(self, tmp_path):
         (tmp_path / ".koan-pause").write_text("1")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["paused"] is True
 
     def test_pause_reason_quota(self, tmp_path):
         (tmp_path / ".koan-pause").write_text("quota\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == "quota"
 
     def test_pause_reason_max_runs(self, tmp_path):
         (tmp_path / ".koan-pause").write_text("max_runs\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == "max_runs"
 
     def test_pause_reason_with_timestamp_line(self, tmp_path):
         """Pause file with 2 lines: reason + unix timestamp."""
         (tmp_path / ".koan-pause").write_text("quota\n1740000000\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path), \
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path), \
              patch("app.reset_parser.time_until_reset", return_value="2h30m"):
-            status = dashboard.get_signal_status()
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == "quota"
             assert "2h30m" in status["reset_time"]
 
@@ -943,15 +944,15 @@ class TestSignalStatusPause:
         (tmp_path / ".koan-pause").write_text(
             "quota\n1740000000\nResets at 15:30\n"
         )
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["reset_time"] == "Resets at 15:30"
 
     def test_pause_reason_bad_timestamp(self, tmp_path):
         """Non-numeric timestamp — should not crash."""
         (tmp_path / ".koan-pause").write_text("quota\nnot-a-number\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == "quota"
             # reset_time stays empty — ValueError caught silently
             assert status["reset_time"] == ""
@@ -959,33 +960,33 @@ class TestSignalStatusPause:
     def test_pause_reason_import_error(self, tmp_path):
         """Missing reset_parser module — should not crash."""
         (tmp_path / ".koan-pause").write_text("quota\n1740000000\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path), \
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path), \
              patch.dict("sys.modules", {"app.reset_parser": None}):
-            status = dashboard.get_signal_status()
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == "quota"
 
     def test_pause_reason_empty_file(self, tmp_path):
         """Empty pause file (legacy touch-created)."""
         (tmp_path / ".koan-pause").write_text("")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["pause_reason"] == ""
 
     def test_daily_report(self, tmp_path):
         (tmp_path / ".koan-daily-report").write_text("5 sessions, 3 productive")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["last_report"] == "5 sessions, 3 productive"
 
     def test_no_daily_report(self, tmp_path):
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert "last_report" not in status
 
     def test_quota_reset_signal(self, tmp_path):
         (tmp_path / ".koan-quota-reset").write_text("1")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
-            status = dashboard.get_signal_status()
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
+            status = dashboard_service.stats.get_signal_status()
             assert status["quota_paused"] is True
 
 
@@ -1000,8 +1001,8 @@ class TestJournalEntries:
         journal = tmp_path / "journal"
         journal.mkdir()
         (journal / "2026-02-15.md").write_text("Flat journal entry.")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries()
             assert len(entries) == 1
             assert entries[0]["date"] == "2026-02-15"
             assert entries[0]["entries"][0]["project"] == "general"
@@ -1014,8 +1015,8 @@ class TestJournalEntries:
         nested = journal / "2026-02-16"
         nested.mkdir()
         (nested / "koan.md").write_text("Nested koan entry.")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries()
             assert len(entries) == 2
             # Most recent first
             assert entries[0]["date"] == "2026-02-16"
@@ -1029,8 +1030,8 @@ class TestJournalEntries:
         nested = journal / "2026-02-15"
         nested.mkdir()
         (nested / "backend.md").write_text("Nested.")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries()
             assert len(entries) == 1
             # Should have both entries for the same date
             assert len(entries[0]["entries"]) == 2
@@ -1042,8 +1043,8 @@ class TestJournalEntries:
             d = journal / f"2026-02-{i+1:02d}"
             d.mkdir()
             (d / "koan.md").write_text(f"Entry {i}")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries(limit=3)
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries(limit=3)
             assert len(entries) == 3
             # Most recent
             assert entries[0]["date"] == "2026-02-10"
@@ -1055,8 +1056,8 @@ class TestJournalEntries:
         (journal / "README.md").write_text("not a date")
         (journal / "2026-02-15").mkdir()
         (journal / "2026-02-15" / "koan.md").write_text("valid")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries()
             assert len(entries) == 1
             assert entries[0]["date"] == "2026-02-15"
 
@@ -1067,8 +1068,8 @@ class TestJournalEntries:
         (d / "koan.md").write_text("Koan work.")
         (d / "backend.md").write_text("Backend work.")
         (d / "tmf.md").write_text("TMF work.")
-        with patch.object(dashboard, "JOURNAL_DIR", journal):
-            entries = dashboard.get_journal_entries()
+        with patch.object(dashboard.state, "JOURNAL_DIR", journal):
+            entries = dashboard_service.journal.get_journal_entries()
             assert len(entries) == 1
             projects = [e["project"] for e in entries[0]["entries"]]
             assert "koan" in projects
@@ -1151,10 +1152,10 @@ class TestReadFile:
     def test_existing_file(self, tmp_path):
         f = tmp_path / "test.md"
         f.write_text("hello world")
-        assert dashboard.read_file(f) == "hello world"
+        assert dashboard_service.read_file(f) == "hello world"
 
     def test_missing_file(self, tmp_path):
-        assert dashboard.read_file(tmp_path / "missing.md") == ""
+        assert dashboard_service.read_file(tmp_path / "missing.md") == ""
 
 
 # ---------------------------------------------------------------------------
@@ -1165,14 +1166,14 @@ class TestBuildDashboardPromptFull:
     """Test _build_dashboard_prompt in full (default) mode."""
 
     def test_full_prompt_includes_journal_and_summary(self, instance_dir):
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-             patch("app.dashboard.load_recent_history", return_value=[]), \
-             patch("app.dashboard.format_conversation_history", return_value=""), \
-             patch("app.dashboard.get_tools_description", return_value=""):
-            prompt = dashboard._build_dashboard_prompt("hello", lite=False)
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+             patch("app.dashboard.chat.load_recent_history", return_value=[]), \
+             patch("app.dashboard.chat.format_conversation_history", return_value=""), \
+             patch("app.dashboard.chat.get_tools_description", return_value=""):
+            prompt = dashboard.chat._build_dashboard_prompt("hello", lite=False)
         assert "You are Kōan" in prompt
         assert "Session 1: bootstrapped" in prompt
 
@@ -1180,14 +1181,14 @@ class TestBuildDashboardPromptFull:
         """Summary > 1500 chars should be truncated."""
         long_summary = "A" * 3000
         (instance_dir / "memory" / "summary.md").write_text(long_summary)
-        with patch.object(dashboard, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
-             patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-             patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-             patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-             patch("app.dashboard.load_recent_history", return_value=[]), \
-             patch("app.dashboard.format_conversation_history", return_value=""), \
-             patch("app.dashboard.get_tools_description", return_value=""):
-            prompt = dashboard._build_dashboard_prompt("hello")
+        with patch.object(dashboard.state, "CONVERSATION_HISTORY_FILE", instance_dir / "history.jsonl"), \
+             patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+             patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+             patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+             patch("app.dashboard.chat.load_recent_history", return_value=[]), \
+             patch("app.dashboard.chat.format_conversation_history", return_value=""), \
+             patch("app.dashboard.chat.get_tools_description", return_value=""):
+            prompt = dashboard.chat._build_dashboard_prompt("hello")
         # Full 3000-char summary should not appear
         assert "A" * 3000 not in prompt
         # But truncated version should
@@ -1220,7 +1221,7 @@ class TestProjectFiltering:
     """Test ?project= filtering on routes and /api/projects."""
 
     def test_api_projects(self, app_client):
-        with patch("app.dashboard.get_known_projects", return_value=[("koan", "/p/koan")]):
+        with patch("app.dashboard_service.missions.get_known_projects", return_value=[("koan", "/p/koan")]):
             resp = app_client.get("/api/projects")
         data = resp.get_json()
         assert "projects" in data
@@ -1228,7 +1229,7 @@ class TestProjectFiltering:
 
     def test_api_projects_includes_mission_tags(self, app_client, instance_dir):
         """Projects from mission tags are included even if not in config."""
-        with patch("app.dashboard.get_known_projects", return_value=[]):
+        with patch("app.dashboard_service.missions.get_known_projects", return_value=[]):
             resp = app_client.get("/api/projects")
         data = resp.get_json()
         # missions.md has [project:koan] tag
@@ -1271,60 +1272,60 @@ class TestProjectFiltering:
             "in_progress": [],
             "done": [],
         }
-        filtered = dashboard._filter_missions_by_project(missions, "koan")
+        filtered = dashboard_service.missions.filter_missions_by_project(missions, "koan")
         assert len(filtered["pending"]) == 1
         assert "Task A" in filtered["pending"][0]
 
     def test_filter_missions_empty_project(self):
         missions = {"pending": ["- Task A", "- Task B"], "in_progress": [], "done": []}
-        filtered = dashboard._filter_missions_by_project(missions, "")
+        filtered = dashboard_service.missions.filter_missions_by_project(missions, "")
         assert filtered == missions
 
     def test_project_badge_filter(self):
-        assert "koan" in dashboard.project_badge_filter("- [project:koan] Fix bug")
-        assert dashboard.project_badge_filter("- Fix bug") == ""
+        assert "koan" in dashboard._helpers.project_badge_filter("- [project:koan] Fix bug")
+        assert dashboard._helpers.project_badge_filter("- Fix bug") == ""
 
     def test_strip_project_tag_filter(self):
-        assert dashboard.strip_project_tag_filter("- [project:koan] Fix bug") == "- Fix bug"
-        assert dashboard.strip_project_tag_filter("- Fix bug") == "- Fix bug"
+        assert dashboard._helpers.strip_project_tag_filter("- [project:koan] Fix bug") == "- Fix bug"
+        assert dashboard._helpers.strip_project_tag_filter("- Fix bug") == "- Fix bug"
 
     def test_linkify_filter_converts_urls(self):
-        result = dashboard.linkify_filter("Fix https://github.com/org/repo/pull/42 now")
+        result = dashboard._helpers.linkify_filter("Fix https://github.com/org/repo/pull/42 now")
         assert 'href="https://github.com/org/repo/pull/42"' in result
         assert 'target="_blank"' in result
         assert 'rel="noopener noreferrer"' in result
 
     def test_linkify_filter_no_url(self):
-        result = dashboard.linkify_filter("Fix the bug")
+        result = dashboard._helpers.linkify_filter("Fix the bug")
         assert "<a " not in result
         assert result == "Fix the bug"
 
     def test_linkify_filter_escapes_html(self):
-        result = dashboard.linkify_filter("<script>alert(1)</script>")
+        result = dashboard._helpers.linkify_filter("<script>alert(1)</script>")
         assert "<script>" not in result
         assert "&lt;script&gt;" in result
 
     def test_linkify_filter_multiple_urls(self):
-        result = dashboard.linkify_filter("See https://jira.example.com/PROJ-123 and https://github.com/org/repo/issues/5")
+        result = dashboard._helpers.linkify_filter("See https://jira.example.com/PROJ-123 and https://github.com/org/repo/issues/5")
         assert result.count("<a ") == 2
 
     def test_linkify_shortens_github_issue_url(self):
-        result = dashboard.linkify_filter("Fix https://github.com/org/repo/issues/123 done")
+        result = dashboard._helpers.linkify_filter("Fix https://github.com/org/repo/issues/123 done")
         assert 'href="https://github.com/org/repo/issues/123"' in result
         assert ">#123</a>" in result
 
     def test_linkify_shortens_github_pull_url(self):
-        result = dashboard.linkify_filter("See https://github.com/org/repo/pull/42")
+        result = dashboard._helpers.linkify_filter("See https://github.com/org/repo/pull/42")
         assert 'href="https://github.com/org/repo/pull/42"' in result
         assert ">#42</a>" in result
 
     def test_linkify_shortens_jira_url(self):
-        result = dashboard.linkify_filter("Fix https://jira.example.com/browse/PROJ-456 now")
+        result = dashboard._helpers.linkify_filter("Fix https://jira.example.com/browse/PROJ-456 now")
         assert 'href="https://jira.example.com/browse/PROJ-456"' in result
         assert ">PROJ-456</a>" in result
 
     def test_linkify_preserves_plain_url(self):
-        result = dashboard.linkify_filter("See https://example.com/docs")
+        result = dashboard._helpers.linkify_filter("See https://example.com/docs")
         assert ">https://example.com/docs</a>" in result
 
 
@@ -1349,7 +1350,7 @@ class TestApiMissionsReorder:
     """Test POST /api/missions/reorder."""
 
     def test_valid_reorder(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/reorder",
                 json={"position": 2, "target": 1})
         assert resp.status_code == 200
@@ -1358,7 +1359,7 @@ class TestApiMissionsReorder:
         assert "pending" in data
 
     def test_invalid_position(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/reorder",
                 json={"position": 99, "target": 1})
         assert resp.status_code == 400
@@ -1370,7 +1371,7 @@ class TestApiMissionsReorder:
         assert resp.status_code == 400
 
     def test_same_position(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/reorder",
                 json={"position": 1, "target": 1})
         assert resp.status_code == 400
@@ -1380,7 +1381,7 @@ class TestApiMissionsCancel:
     """Test POST /api/missions/cancel."""
 
     def test_valid_cancel(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/cancel",
                 json={"position": 1})
         assert resp.status_code == 200
@@ -1390,7 +1391,7 @@ class TestApiMissionsCancel:
         assert "pending" in data
 
     def test_invalid_position(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/cancel",
                 json={"position": 99})
         assert resp.status_code == 400
@@ -1404,7 +1405,7 @@ class TestApiMissionsEdit:
     """Test POST /api/missions/edit."""
 
     def test_valid_edit(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/edit",
                 json={"position": 1, "text": "Updated mission"})
         assert resp.status_code == 200
@@ -1414,13 +1415,13 @@ class TestApiMissionsEdit:
         assert "Updated mission" in content
 
     def test_empty_text(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/edit",
                 json={"position": 1, "text": ""})
         assert resp.status_code == 400
 
     def test_invalid_position(self, app_client, instance_dir):
-        with patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"):
+        with patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"):
             resp = app_client.post("/api/missions/edit",
                 json={"position": 99, "text": "New text"})
         assert resp.status_code == 400
@@ -1474,29 +1475,29 @@ Done — findings documented
 """
 
     def test_extracts_phases(self):
-        result = dashboard._parse_plan_progress(self._STANDARD_PLAN)
+        result = dashboard_service.plans.parse_plan_progress(self._STANDARD_PLAN)
         assert result["total"] == 3
         assert result["phases"][0]["title"] == "Backend endpoints"
         assert result["phases"][1]["title"] == "Frontend"
         assert result["phases"][2]["title"] == "Tests"
 
     def test_detects_checkmark_completion(self):
-        result = dashboard._parse_plan_progress(self._COMPLETED_PHASE)
+        result = dashboard_service.plans.parse_plan_progress(self._COMPLETED_PHASE)
         assert result["total"] == 1
         assert result["completed"] == 1
         assert result["phases"][0]["completed"] is True
         assert result["percent"] == 100
 
     def test_detects_checkbox_completion(self):
-        result = dashboard._parse_plan_progress(self._CHECKBOX_PHASE)
+        result = dashboard_service.plans.parse_plan_progress(self._CHECKBOX_PHASE)
         assert result["phases"][0]["completed"] is True
 
     def test_detects_done_text_completion(self):
-        result = dashboard._parse_plan_progress(self._DONE_TEXT_PHASE)
+        result = dashboard_service.plans.parse_plan_progress(self._DONE_TEXT_PHASE)
         assert result["phases"][0]["completed"] is True
 
     def test_incomplete_phases(self):
-        result = dashboard._parse_plan_progress(self._STANDARD_PLAN)
+        result = dashboard_service.plans.parse_plan_progress(self._STANDARD_PLAN)
         # Phase 1 and 2 have no completion markers; Phase 3 has ✅
         assert result["phases"][0]["completed"] is False
         assert result["phases"][1]["completed"] is False
@@ -1505,17 +1506,17 @@ Done — findings documented
         assert result["percent"] == 33
 
     def test_empty_markdown(self):
-        result = dashboard._parse_plan_progress("")
+        result = dashboard_service.plans.parse_plan_progress("")
         assert result == {"phases": [], "completed": 0, "total": 0, "percent": 0}
 
     def test_no_phases(self):
-        result = dashboard._parse_plan_progress("# Some title\n\nNo phases here.")
+        result = dashboard_service.plans.parse_plan_progress("# Some title\n\nNo phases here.")
         assert result["total"] == 0
         assert result["percent"] == 0
 
     def test_malformed_plan_best_effort(self):
         """Plans that don't follow the strict format return gracefully."""
-        result = dashboard._parse_plan_progress("Random content\nwithout phases")
+        result = dashboard_service.plans.parse_plan_progress("Random content\nwithout phases")
         assert result["total"] == 0
 
 
@@ -1541,7 +1542,7 @@ class TestPlansPage:
 
     def test_api_plans_skips_projects_without_github_url(self, app_client):
         with patch("app.utils.get_known_projects", return_value=[("myproject", "/some/path")]), \
-             patch("app.dashboard._get_project_repo", return_value=None):
+             patch("app.dashboard_service.plans.get_project_repo", return_value=None):
             resp = app_client.get("/api/plans")
         assert resp.status_code == 200
         data = resp.get_json()
@@ -1557,8 +1558,8 @@ class TestPlansPage:
             "url": "https://github.com/owner/repo/issues/42",
         }])
         with patch("app.utils.get_known_projects", return_value=[("myproject", "/path")]), \
-             patch("app.dashboard._get_project_repo", return_value="owner/repo"), \
-             patch.dict("app.dashboard._plans_cache", {}, clear=True), \
+             patch("app.dashboard_service.plans.get_project_repo", return_value="owner/repo"), \
+             patch.dict("app.dashboard.state._plans_cache", {}, clear=True), \
              patch("app.github.run_gh", return_value=gh_response):
             resp = app_client.get("/api/plans")
         assert resp.status_code == 200
@@ -1575,8 +1576,8 @@ class TestPlansPage:
         """Project filter limits results to matching project."""
         with patch("app.utils.get_known_projects",
                    return_value=[("proj_a", "/a"), ("proj_b", "/b")]), \
-             patch("app.dashboard._get_project_repo", return_value="owner/repo"), \
-             patch.dict("app.dashboard._plans_cache", {}, clear=True), \
+             patch("app.dashboard_service.plans.get_project_repo", return_value="owner/repo"), \
+             patch.dict("app.dashboard.state._plans_cache", {}, clear=True), \
              patch("app.github.run_gh", return_value="[]"):
             resp = app_client.get("/api/plans?project=proj_a")
         assert resp.status_code == 200
@@ -1597,8 +1598,8 @@ class TestPlansPage:
             "body": "", "updatedAt": "", "url": "",
         }])
         with patch("app.utils.get_known_projects", return_value=[("myproject", "/p")]), \
-             patch("app.dashboard._get_project_repo", return_value="owner/repo"), \
-             patch.dict("app.dashboard._plans_cache", fresh_cache, clear=True), \
+             patch("app.dashboard_service.plans.get_project_repo", return_value="owner/repo"), \
+             patch.dict("app.dashboard.state._plans_cache", fresh_cache, clear=True), \
              patch("app.github.run_gh", return_value=fresh_gh):
             # Without force — should use cache
             resp = app_client.get("/api/plans")
@@ -1611,7 +1612,7 @@ class TestPlansPage:
             assert data["plans"][0]["title"] == "Fresh"
 
     def test_api_plan_detail_no_github_url(self, app_client):
-        with patch("app.dashboard._get_project_repo", return_value=None):
+        with patch("app.dashboard_service.plans.get_project_repo", return_value=None):
             resp = app_client.get("/api/plans/myproject/42")
         assert resp.status_code == 404
 
@@ -1627,9 +1628,9 @@ class TestPlansPage:
                 {"body": "#### Phase 1: Setup\n\n✅ Done.", "createdAt": "2026-03-14T11:00:00Z"}
             ],
         })
-        with patch("app.dashboard._get_project_repo", return_value="owner/repo"), \
+        with patch("app.dashboard_service.plans.get_project_repo", return_value="owner/repo"), \
              patch("app.github.run_gh", return_value=gh_response), \
-             patch.object(dashboard, "MISSIONS_FILE", Path("/nonexistent/missions.md")):
+             patch.object(dashboard.state, "MISSIONS_FILE", Path("/nonexistent/missions.md")):
             resp = app_client.get("/api/plans/myproject/42")
         assert resp.status_code == 200
         data = resp.get_json()
@@ -1648,8 +1649,8 @@ class TestPlansPage:
             "- /plan https://github.com/owner/repo/issues/42\n"
             "- Some unrelated mission\n"
         )
-        with patch.object(dashboard, "MISSIONS_FILE", missions_file):
-            linked = dashboard._find_linked_missions(
+        with patch.object(dashboard.state, "MISSIONS_FILE", missions_file):
+            linked = dashboard_service.plans.find_linked_missions(
                 "https://github.com/owner/repo/issues/42", 42
             )
         assert len(linked) == 1
@@ -1661,15 +1662,15 @@ class TestPlansPage:
             "myproj": {"github_url": "https://github.com/owner/repo"}
         }}
         with patch("app.projects_config.load_projects_config", return_value=fake_cfg), \
-             patch.object(dashboard, "KOAN_ROOT", Path("/fake/root")):
-            result = dashboard._get_project_repo("myproj")
+             patch.object(dashboard.state, "KOAN_ROOT", Path("/fake/root")):
+            result = dashboard_service.plans.get_project_repo("myproj")
         assert result == "owner/repo"
 
     def test_get_project_repo_returns_none_when_no_config(self):
         """_get_project_repo returns None when projects.yaml can't be loaded."""
         with patch("app.projects_config.load_projects_config", return_value=None), \
-             patch.object(dashboard, "KOAN_ROOT", Path("/fake/root")):
-            result = dashboard._get_project_repo("myproj")
+             patch.object(dashboard.state, "KOAN_ROOT", Path("/fake/root")):
+            result = dashboard_service.plans.get_project_repo("myproj")
         assert result is None
 
 
@@ -1684,13 +1685,13 @@ class TestRulesRoutes:
     """Integration tests for the /api/rules and /rules endpoints."""
 
     def test_get_rules_empty(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.get("/api/rules")
         assert resp.status_code == 200
         assert resp.get_json() == []
 
     def test_post_rule_creates_entry(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.post("/api/rules", json={
                 "event": "post_mission",
                 "action": "notify",
@@ -1710,7 +1711,7 @@ class TestRulesRoutes:
             assert rules[0]["id"] == rule["id"]
 
     def test_post_rule_unknown_event_returns_400(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.post("/api/rules", json={
                 "event": "no_such_event",
                 "action": "notify",
@@ -1719,7 +1720,7 @@ class TestRulesRoutes:
         assert "error" in resp.get_json()
 
     def test_post_rule_unknown_action_returns_400(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.post("/api/rules", json={
                 "event": "post_mission",
                 "action": "send_email",
@@ -1727,7 +1728,7 @@ class TestRulesRoutes:
         assert resp.status_code == 400
 
     def test_patch_rule_toggles_enabled(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             create = app_client.post("/api/rules", json={
                 "event": "post_mission",
                 "action": "notify",
@@ -1741,7 +1742,7 @@ class TestRulesRoutes:
             assert patch_resp.get_json()["enabled"] is False
 
     def test_delete_rule_removes_it(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             create = app_client.post("/api/rules", json={
                 "event": "pre_mission",
                 "action": "pause",
@@ -1755,20 +1756,20 @@ class TestRulesRoutes:
             assert all(r["id"] != rule_id for r in rules)
 
     def test_delete_nonexistent_rule_returns_404(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.delete("/api/rules/does_not_exist")
         assert resp.status_code == 404
 
     def test_rules_page_renders_without_error(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.get("/rules")
         assert resp.status_code == 200
         assert b"Automation Rules" in resp.data
 
     def test_rules_page_shows_empty_state_when_no_rules(self, app_client, instance_dir):
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-             patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+             patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"):
             resp = app_client.get("/rules")
         assert resp.status_code == 200
         assert b"No rules yet" in resp.data
@@ -1779,7 +1780,7 @@ class TestApiLogs:
 
     def test_api_logs_no_log_dir(self, app_client, tmp_path):
         """Returns empty lines when $KOAN_ROOT/logs/ doesn't exist."""
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs")
         assert resp.status_code == 200
         data = resp.get_json()
@@ -1791,7 +1792,7 @@ class TestApiLogs:
         logs_dir = tmp_path / "logs"
         logs_dir.mkdir()
         (logs_dir / "run.log").write_text("line1\nline2\nline3\n")
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs?source=run")
         assert resp.status_code == 200
         data = resp.get_json()
@@ -1806,7 +1807,7 @@ class TestApiLogs:
         (logs_dir / "run.log").write_text(
             "info: all good\nerror: something failed\ninfo: still ok\n"
         )
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs?source=run&q=error")
         data = resp.get_json()
         assert data["total"] == 1
@@ -1818,7 +1819,7 @@ class TestApiLogs:
         logs_dir.mkdir()
         content = "\n".join(f"line{i}" for i in range(300)) + "\n"
         (logs_dir / "run.log").write_text(content)
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs?source=run&limit=50")
         data = resp.get_json()
         assert data["total"] == 50
@@ -1829,7 +1830,7 @@ class TestApiLogs:
         logs_dir.mkdir()
         content = "\n".join(f"line{i}" for i in range(10)) + "\n"
         (logs_dir / "run.log").write_text(content)
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs?source=run&limit=99999")
         assert resp.status_code == 200  # no error
 
@@ -1839,7 +1840,7 @@ class TestApiLogs:
         logs_dir.mkdir()
         (logs_dir / "run.log").write_text("run line\n")
         # awake.log intentionally absent
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/logs?source=all")
         data = resp.get_json()
         sources = {e["source"] for e in data["lines"]}
@@ -1852,7 +1853,7 @@ class TestApiHealth:
 
     def test_api_health_structure(self, app_client, tmp_path):
         """Response contains disk, run, and awake keys."""
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/health")
         assert resp.status_code == 200
         data = resp.get_json()
@@ -1862,7 +1863,7 @@ class TestApiHealth:
 
     def test_api_health_no_pids(self, app_client, tmp_path):
         """With no PID files, run.alive and awake.alive are False."""
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path):
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
             resp = app_client.get("/api/health")
         data = resp.get_json()
         assert data["run"]["alive"] is False
@@ -1874,8 +1875,8 @@ class TestApiHealth:
         used = 90 * 1024 * 1024
         free = total - used
         mock_usage = shutil.disk_usage.__class__  # just need the namedtuple shape
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path), \
-             patch("app.dashboard.shutil.disk_usage",
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path), \
+             patch("app.dashboard.core.shutil.disk_usage",
                    return_value=type("DiskUsage", (), {"total": total, "used": used, "free": free})()):
             resp = app_client.get("/api/health")
         data = resp.get_json()
@@ -1887,8 +1888,8 @@ class TestApiHealth:
         total = 100 * 1024 * 1024
         used = 96 * 1024 * 1024
         free = total - used
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path), \
-             patch("app.dashboard.shutil.disk_usage",
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path), \
+             patch("app.dashboard.core.shutil.disk_usage",
                    return_value=type("DiskUsage", (), {"total": total, "used": used, "free": free})()):
             resp = app_client.get("/api/health")
         data = resp.get_json()
@@ -1899,8 +1900,8 @@ class TestApiHealth:
         total = 100 * 1024 * 1024
         used = 50 * 1024 * 1024
         free = total - used
-        with patch.object(dashboard, "KOAN_ROOT", tmp_path), \
-             patch("app.dashboard.shutil.disk_usage",
+        with patch.object(dashboard.state, "KOAN_ROOT", tmp_path), \
+             patch("app.dashboard.core.shutil.disk_usage",
                    return_value=type("DiskUsage", (), {"total": total, "used": used, "free": free})()):
             resp = app_client.get("/api/health")
         data = resp.get_json()
@@ -2063,7 +2064,7 @@ class TestAgentControls:
 
     def test_pause_with_duration(self, app_client, tmp_path):
         fixed_time = 1700000000
-        with patch("app.dashboard.time.time", return_value=fixed_time):
+        with patch("app.dashboard.prs.time.time", return_value=fixed_time):
             resp = app_client.post("/api/agent/pause", json={"duration": "2h"})
         data = resp.get_json()
         assert resp.status_code == 200
@@ -2138,7 +2139,7 @@ class TestNicknameApi:
         import yaml
         config_path = instance_dir / "config.yaml"
         config_path.write_text("dashboard:\n  enabled: true\n")
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.put("/api/nickname",
                 json={"nickname": "Prod-1"},
                 content_type="application/json")
@@ -2152,7 +2153,7 @@ class TestNicknameApi:
     def test_set_nickname_truncates_at_50(self, app_client, instance_dir):
         config_path = instance_dir / "config.yaml"
         config_path.write_text("")
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.put("/api/nickname",
                 json={"nickname": "x" * 100},
                 content_type="application/json")
@@ -2163,7 +2164,7 @@ class TestNicknameApi:
         import yaml
         config_path = instance_dir / "config.yaml"
         config_path.write_text("dashboard:\n  nickname: OldName\n")
-        with patch.object(dashboard, "INSTANCE_DIR", instance_dir):
+        with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir):
             resp = app_client.put("/api/nickname",
                 json={"nickname": ""},
                 content_type="application/json")
@@ -2328,35 +2329,35 @@ class TestPassphraseGate:
     """The dashboard passphrase gate (KOAN_DASHBOARD_PWD)."""
 
     def test_no_passphrase_allows_access(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", ""):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", ""):
             resp = app_client.get("/api/status")
             assert resp.status_code == 200
 
     def test_passphrase_blocks_html_with_redirect(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             resp = app_client.get("/")
             assert resp.status_code == 302
             assert "/login" in resp.headers["Location"]
 
     def test_passphrase_blocks_api_with_401(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             resp = app_client.get("/api/status")
             assert resp.status_code == 401
 
     def test_login_page_renders(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             resp = app_client.get("/login")
             assert resp.status_code == 200
             assert b"Passphrase" in resp.data
 
     def test_wrong_passphrase_rejected(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             resp = app_client.post("/login", data={"passphrase": "nope"})
             assert resp.status_code == 200
             assert b"Incorrect passphrase" in resp.data
 
     def test_correct_passphrase_unlocks_session(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             resp = app_client.post("/login", data={"passphrase": "s3cret"})
             assert resp.status_code == 302
             # Now an authenticated request goes through.
@@ -2364,7 +2365,7 @@ class TestPassphraseGate:
             assert resp2.status_code == 200
 
     def test_logout_clears_session(self, app_client):
-        with patch.object(dashboard, "DASHBOARD_PWD", "s3cret"):
+        with patch.object(dashboard.state, "DASHBOARD_PWD", "s3cret"):
             app_client.post("/login", data={"passphrase": "s3cret"})
             app_client.get("/logout")
             resp = app_client.get("/api/status")
