+++ koan/tests/test_pr_tracker.py
@@ -12,7 +12,7 @@ from jinja2 import FileSystemLoader
 
 from app import pr_tracker, dashboard
 
-REAL_TEMPLATES = Path(__file__).parent.parent / "templates"
+REAL_TEMPLATES = Path(__file__).parent.parent / "templates" / "dashboard"
 
 
 # ---------------------------------------------------------------------------
@@ -322,14 +322,14 @@ def app_client(instance_dir, tmp_path):
     """Create a Flask test client with patched paths."""
     tpl_dest = tmp_path / "koan" / "templates"
     shutil.copytree(REAL_TEMPLATES, tpl_dest)
-    with patch.object(dashboard, "INSTANCE_DIR", instance_dir), \
-         patch.object(dashboard, "MISSIONS_FILE", instance_dir / "missions.md"), \
-         patch.object(dashboard, "OUTBOX_FILE", instance_dir / "outbox.md"), \
-         patch.object(dashboard, "SOUL_FILE", instance_dir / "soul.md"), \
-         patch.object(dashboard, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
-         patch.object(dashboard, "JOURNAL_DIR", instance_dir / "journal"), \
-         patch.object(dashboard, "PENDING_FILE", instance_dir / "journal" / "pending.md"), \
-         patch.object(dashboard, "KOAN_ROOT", tmp_path):
+    with patch.object(dashboard.state, "INSTANCE_DIR", instance_dir), \
+         patch.object(dashboard.state, "MISSIONS_FILE", instance_dir / "missions.md"), \
+         patch.object(dashboard.state, "OUTBOX_FILE", instance_dir / "outbox.md"), \
+         patch.object(dashboard.state, "SOUL_FILE", instance_dir / "soul.md"), \
+         patch.object(dashboard.state, "SUMMARY_FILE", instance_dir / "memory" / "summary.md"), \
+         patch.object(dashboard.state, "JOURNAL_DIR", instance_dir / "journal"), \
+         patch.object(dashboard.state, "PENDING_FILE", instance_dir / "journal" / "pending.md"), \
+         patch.object(dashboard.state, "KOAN_ROOT", tmp_path):
         dashboard.app.config["TESTING"] = True
         dashboard.app.jinja_loader = FileSystemLoader(str(tpl_dest))
         with dashboard.app.test_client() as client:
